pyproject.toml
[project]
name = "cerberus-dpl"
version = "0.1.0"
description = "Cerberus-DPL is a lightweight pipeline for filtering, normalizing and ingesting documents into RAG systems."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = "Apache-2.0"
dependencies = [
    "bertopic>=0.17.3",
    "docling>=1.11.0",
    "hdbscan>=0.8.33",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.0",
    "structlog>=25.4.0",
    "typer>=0.17.4",
    "umap-learn>=0.5.4",
]

[project.scripts]
cerberus = "cerberus_dpl.cli:main"

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "ruff>=0.12.12",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
